+++ CHANGELOG.md
@@ -6,3 +6,14 @@
 ## 0.0.2
 
 * TODO: New function updated : function to delete all notification 
+
+
+## 0.0.3
+
+* TODO: New function updated : payload data field added in notification model 
+
+
+## 0.0.4
+
+* New function added: Function to delete notifications.
+* TODO: Function `markAsRead` is deprecated. Use the new method `markAllAsRead` instead.